pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-lessons"
version = "0.1.0"
description = "AI-powered lessons learned system with semantic search and graph relationships"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Guy Mitchell"}
]
keywords = ["ai", "lessons", "knowledge-base", "semantic-search", "mcp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "python-ulid>=2.0",
    "sqlite-vec>=0.1.1",
    "pysqlite3-binary>=0.5.0",
    "sentence-transformers>=2.2.0",
]

[project.optional-dependencies]
openai = [
    "openai>=1.0",
]
mcp = [
    "mcp>=1.0.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]
all = [
    "ai-lessons[openai,mcp,dev]",
]

[project.scripts]
ai-lessons = "ai_lessons.cli:main"
ai-lessons-mcp = "ai_lessons.mcp_server:run"

[project.urls]
Homepage = "https://github.com/shitchell/ai-lessons"
Repository = "https://github.com/shitchell/ai-lessons"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
